# Improve performance of json functions